requirements/base.txt
python-decouple==3.4
structlog==21.1.0
orjson==3.6.4
